Add an optional paramiko dependency
It's useful so that I can use the --host option to artiq_flash.py. I've made it
optional I think you'll need to run "poetry install -E remote-flashing" to
install it but I've not been able to test that because it was already for me when
I did the poetry add.