Improve error reporting in some tests
When an exception was thrown in the worker process started by these
tests this wasn't visible for two reasons.

The exception isn't propagated in to the tests so they would just sit
there and hang. When you interrupted the tests you didn't see any
output from them. With this fixed you can see that an exception was
thrown but not what it is.

The worker process was configured not to log anything below critical
and the exception was logged as error.