Factor out `PipeWorkerTransport`
There's something interesting going on with `io_lock`. I think that
it's just really to ensure that we write whole lines to the pipe
uninterleaved. I wanted to move the io_lock in to the transport
object because I think that the requirements for the multiplexed
implementation are going to be quite different here.

I'm concerned that there's some interaction with the
closed event, previously we were only checking the closed event with
the lock held. But I don't think that's necessary. We preserve the
waiting behaviour whilst io_lock is held so either the closed event is
already set, so it doesn't matter, or we set the event whilst it's
trying to do io in which case we cancel the io. Seems fine.

Generally I think it's quite nice to have kept as much as possible
of the timeout logic in `Worker` because I think that'll be shared
between the transports.