worker transport: Remove the io_lock from recv
Otherwise we get a deadlock if we need to send something
before the next recv is expected. The WorkerManager is less aware
of the state of the worker and is always trying to recv so this
means that send always deadlocks.

This is fine in the `Worker` case because we always do a recv
in a wait for or cancel with the close event so when we start close
any outstanding recv will be cancelled.

In `WorkerManager` we will have to stop the recv task when we want
to close a worker.