Launch worker manager from artiq client on request.
Pass --start-worker-mgr to have artiq client start a worker manager
and use that to run the experiment. Log output is just forwarded to
stdout of the parent process.
artiq_worker_manager get's a new --exit-on-idle flag, which means that it
will exit when no experiments are running (having run at least one). This
might not work well with periodic experiments.