Fix error reporting if the experiment fails to import
Ensure that we signal that the worker has failed if it exits and that
that causes a recv to return an empty string as it does for the pipe
worker.
Handle overflow of stderr/stdout queues. Not super happy about this
especially as the task that consumes this doesn't do any async. But
we need this interface for the task that consumes this in worker
manager where we do some async.