Propagate errors with in the worker manager
This is an attempt to centralise error handling. Rather than
have each task to error handling keep track of all the tasks
we create in a master task. I think this is like structured
concurrency. I can kind of see what people are getting out of
things like curio/trio now.

This always brings down the entire worker manager, but I think
that's ok. If I were to add failing individual workers then I
think I'd do that by having a WorkerError exception with a
worker_id that we could raise from anywhere and would be
handled in the main loop by creating a new task to close down
that worker.