Add ThreadWorkerTransport
This is an alternative Worker transport that runs the
worker in the current process meaning that you can continue
interactive debugger sessions in to the experiment.

It relies on a partial reimplementation of sipyco's pipe_ipc.
In order to keep the changes to worker_impl to be reasonably
minimal I didn't want to have to inject and alternate
communication path. But sipyco's pipe_ipc assumes that the
pipe is going to be read from another process.