Store experiment results on the master
The build step changes the working directory so that any files
created by the experiment would be created in the results
directory. I'm a bit worried that people depend on this to write
ad-hoc results files from inside their experiments. So I've left
that behaviour as is.

I've changed the storage of the results to always store them
according to the UTC time of the experiment start rather than
the local time. There's no guarantees that the server that the
master is running on will have a sensible timezone configured.

This required a change to the analyze step, so that it wrote the
results before sending complete. Otherwise, there was a race with
the master deleting the worker and we frequently (100% of 1 trial)
saw the "terminate" request in response to store_results request.
Rather than seeing the store_results response first.

Also remove the locking from PipeWorkerTransport and move the
setting of the closed event until after the worker is closed.
This allows us to make an RPC call from the worker to the master
after the master has started closing the worker, which occurs if
the worker was in the idle state (during run) so that the
experiment gets terminated midway through. This lock would
protect us if the send or the recieve were not atomic but they
are. It also prevents other operations during close but we
actively want these now. And I don't think it will be
problematic to do them. With this lock we risk deadlocks which
are worse.