Support scanning using worker managers
Adds some regression tests as well as tests for scanning repos
associated with the worker managers.
Worker managers now have a repo root defaulting to the working
directory.
The ExperimentDB has an extra layer of indirection. An
experiment repo that's builtin (id=None) and one for each
worker manager. This is created lazily when that worker manager
is accessed.
We do all the worker transport creation through the worker
manager now. So there's no need to pass the worker manager db in
to the scheduler.
The ExperimentDB exposes local_explist and local_status which are
backwards compatible with the existing explist and explist_status
entries in the publiser. The new all_explist and all_explist_status
publisher targets contain the existing stuff keyed by None and the
worker manager explists keyed by worker manager id.