Fix worker deletion
When we added sending results to the master it became possible that
the worker would want to send and recv during the close stage. I
allowed this by setting the close event after doing the transport
close. This was bad because it meant that errors that were propagated
out of worker actions due to the close got out before closed was set
meaning that they didn't get filtered by: artiq/master/scheduler.py:44
which in turn meant that the caused delete_cb to be called again by
RunStage, PrepareStage, etc. which could fail and leave the stage in
a broken state that prevented the deletion of the pipeline, which
could prevent any further experiments from running.

This change allows the worker request handler, recv and send all to
continue operating after close has been set. But stops the worker
action. So that we can set close at the start of the closing sequence.
To enable this the worker request handler is its own task, which runs
continuously whilst the worker is running. Communication with this
task is achieved by a future, which makes reporting errors a bit more
verbose and draws attention to some edge cases that already existed
as well as introducing some more that needed to be handled. But it
means that we'll always be sensitive to worker exit so this will be
reported to the logs in a timely fashion.