Fix sending responses to worker after delete
Further problems with the "send results code". If the
experiment exits because of a delete on the dashboard then the
experiment will get removed from the workers map on in the
worker manager before the results are sent. Fix that.

Also, we would get out of step in the request and response
pattern between the worker and Worker._handle_worker_requests
because one of the worker requests will not have consumed its
response and consumed the terminate message instead. So we need
to resynchronise. Which we do my remembering that we expect an
extra status and discarding it. We're unlikely to notice if we
get this wrong see and the only requests that the worker sends
after receiving a terminate are store_result requests, which
don't have a response body.