Remove forwarding of RecursionError
This was copied in from the standard library's StreamHandler.emit[0],
to worker around https://bugs.python.org/issue36272.
But isn't present in any of the other standard library Handler
subclasses. The difference in behaviour it achieves seems quite
marginal. Exit with abort vs. exit with RecursionError. And the
chances of hitting this seem quite slim. So I think we can do with out
it.

[0] https://github.com/python/cpython/blob/9b51fd5d/Lib/logging/__init__.py#L1109