Orderly shutdown of TCP sockets
When adding a test that duplicate connections are rejected, I
discovered that they were, but with a reset.

This change makes the shutdown of the TCP connections between
the artiq master and the worker manager more orderly. i.e. we
shutdown write then read any warn about any messages that come
in that we're not going to process before closing the socket.
This should ensure that we don't get any resets on the wire unless
there's been a network error of some kind. We also shouldn't get
so many spurious connection reset back traces.

We only do the orderly shutdown at the end that initiates closing
the connection in the case that there's not been a network error
already.