Garbage collect worker manager proxies
Worker managers now request a period of time for the master to keep
their state alive. If the worker manager doesn't reconnect in that
period then the state is disposed of by the master.
We set the reconnect-timeout in the worker manager depending on
whether this is a real worker manager that's been given an ID
(a pattern which is matched by the ones started by the dashboard),
in which case we request a long reconnect-timeout, otherwise we
don't need to reconnect and so we ask for a 0 reconnect-timeout.