Reduce calls to get_device during compile
When visiting a function call the compiler wants to verify that the function
is intended for the same core device. We have a core device object but not a
name, so we get the core device object from the targeted core device name of
the function. Previously we did that for every function call. We now cache a
successful result so that we don't need to repeat the get_device call after
we've found the correct name.