Add dumb cache to ExamineDatasetMgr.get
When we're doing examine operations it doesn't matter too much if we have
slightly out of date datasets. They'll be used to populate argument defaults
but not for the actual experiments.

The worker processes used for examine are quite short lived so we can use
a pretty dumb cache. This speeds up examines and scan where lots of
experiments use the same datasets. Especially if we're in sandboxing mode
and there's a ropey network hop.