dashboard: Remember the currently selected repo
When I move the canonical experiments from built-in to default everyone
is going to complain that they have to remember to change the repo
everytime the dashboard starts. At least with this change they'll only
have to do it once, and it will be remembered. Also, if you leave it
on local then that'll be remembered too.