No drop log forwarding
Use an async callback instead of generating an async iterator for log
forwarding. This means that there is no longer any need for application
level queues. Back pressure is applied all the way back to the experiment
meaning that if logging can't be processed fast enough then the experiment
will be slowed down. For real experiments using the `PipeWorkerTransport`
there's effectively a large buffer in the stdout and stderr pipes meaning
that this shouldn't generally be a problem.

I feel a bit unhappy that there's a hidden preference on the behaviour of
the `OutputHandler` call back received by the `ManagedWorkerTransport`.
We don't really want this to actually await internally and block the
receive-loop in the `WorkerManagerProxy`. But it's always provided the
callback that just writes the log line directly to the master log and that's
unlikely to change so that's fine.