Add dataset cache during experiments
The cache is cleared between each action and when ever we're paused.
The idea is that most the actions should be short enough that datasets
changing during them should be rare and not matter that much. Except
run, but we assume that datasets are mostly updated in run, and we're
the only experiment running so they shouldn't be being updated.
Unless we pause, but then we clear the cache.