Allow updating the log filter retrospectively
We now store the past 10k log messages in memory (rather than the past
1k matching ones). This allows us to refilter this list when the user
changes filters (log level, text filter and worker managers).

Artiq used to support this many years ago but it was removed [^1]. This
patch is almost a direct revert of that commit with some minor
performance improvements[^2].

It's unclear what was wrong with the previous implementation, but worth
bearing in mind.

[^1]: https://github.com/m-labs/artiq/commit/10267f39c963b60d0b085b2c6e05498fabfea946

[^2]: The viewer is now able to navigate 10k log entries and refilter it
with tolerable (~1s delay) latencies.