Fix race in initialising worker manager list
On Windows its possible for the initial worker manager list to be
loaded before we have figured out the local worker manager's id. We might
then order the local worker manager incorrectly. If we have and we later
move another row in the worker managers table we can issue invalid
startMoveRows/endMoveRows calls because the rows aren't in the order they're
meant to be according to the sort key. I think this is a source of crashes
in the endMoveRows function.

To avoid this we don't attempt to show any of the worker managers until we
have started the local worker manager