Pass correct new row to beginMoveRows
beginMoveRows wants the target row index as if the row has not been moved yet.
When we insert in to row_to_key, we want to use the row index as if it had
been moved. Because we've already removed it. Move the adjustment of the row
index for the row moving until after we've called `beginMoveRows`.

Closes #162