[WIP] Add eye scan functionality
Attempt to add eye scan functionality into artiq to allow us to debug transceiver / DRTIO issues.

Over the weekend @cjbe run an eye scan on some hardware at home using the standard xilinx cores. The outcomes looked really good, suggesting there is nothing grossly broken with our hardware (good!). This test uses the "default" configuration for the transceivers and doesn't touch any artiq code / settings.

![image](https://github.com/OxIonics/artiq/assets/21218399/faa4842f-d3e2-4496-9c29-5928ddf7f74c)

@cjbe what's the best reference for this? Is it the [transceiver user guide](https://docs.amd.com/v/u/en-US/ug476_7Series_Transceivers) RX margin analysis section (pp 214 onwards). Also [UG482](https://docs.amd.com/v/u/en-US/ug482_7Series_GTP_Transceivers)

[Example code](https://support.xilinx.com/s/article/64098?language=en_US)

The next step would be to integrate this with artiq to allow us to get signal integrity data out of live hardware. This would, for example, show up if the transceivers are misconfured in a way that's leading to SI problems.

As far as I understand it, @cjbe has had an initial go at this (this PR) but it's not giving reproducible results.

Reviewers:

Pull Request: https://github.com/OxIonics/artiq/pull/350